fix(rtc): honor negotiated media direction
